pyproject.toml
[tool.poetry]
name = "arealglam-gpt-backend"
version = "0.1.0"
description = "arealglam python backend for OpenAI gpt integration"
authors = ["Kevin <kevin.armengol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.0"
openai = "^1.3.6"
instructor = "0.3.4"
python-dotenv = "1.0.0"
deepdiff = "6.7.1"
termcolor = "2.3.0"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
jupytext = "^1.15.2"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
